Prevent rejection of `data.get()`
If `data.reload()` were to reject, `dataBlockingPromise` and therefore
`data.get()` would reject as well. To counteract that, if the promise
retrieving the data rejects, the promise will still resolve as the old
data. This keeps things easy. Now, if an error happens on initial
startup, that may be a problem.