Reallow rejection of `data.get()`
After further thought, it was a good idea to allow rejection, cause that would
then enable the "unable to fetch" error messages. A problem is by caching the
return value of `data.get()` (as I sometimes did), I am concerned
that the cache value may be bad by the time some of the promises that use
this cached value resolve. I'll just kick the can down the road.