Handle messages that are over character limit
Discord has a character limit in messages, so big messages must be
chunked up into multiple messages. A new function was added to do all
of this.

This commit also creates a new file, `util.js`. This file is for
utility functions that don't depend on anything other than the
arguments passed to the function.