Combine map code and data code into season code
In the code, the data and the map code have been treated separately:
I have gone out of my way to keep these separate, but I am realizing
more and more that this is unnecessary.

This commit merges the season data and map into one file and cleans up
the process of saving these files.